(fix) Speed up page turning by using org-element-cache-map.
This commit focuses on the org-noter--get-view-info function,
a function that help org-noter decides the best heading to serve
as a "reference" heading.

Previously, the function uses org-element-map which is quite slow
and should only be used when buffer modification is needed, which
in this case, it isn't.

I made two changes to this function.
First, use the more performant variant of org-element-map,
org-element-cache-map.
Second, make the function works when the root element is a file-level
property drawer.

Since org-noter--get-view-info works by mapping all elements in the
note buffer every time user changes a page; consequently, when user
have a really large note with a lot of headings each referencing some
pages in the document, the user has to wait for the function to map
all the headings before the page is turned.