(fix) Not parsing the document prop before checking if it's root
This happens when, for example, the document property is a citekey
like this: "cite:&johndoe2020" and org-noter will parse this to get
the file name of the document and store this:
"/path/to/johndoe2020.pdf" as the file that we wanted.

Without parsing, org-noter will be matching "cite:&johndoe2020" to
"/path/to/johndoe2020.pdf" and give false results.