fix sync-next-page-or-chapter navigation
GS injects `most-postitive-fixnum' as the precise vertical location of a page
when searching forward, so -approx-location-cons functions MUST be able to
handle (page . v) type locations.  This was broken at Ahmed Shariff's 3rd
commit.

Also, in "next" navigation, org-element-map can stop as soon as it finds a
match, so the "nil" in the call to macro
`org-noter--map-ignore-headings-with-doc-file' works with "t" and avoids extra
iterations of org-element-map.