-bug (navigation scrolling) +feature (vscroll buffer)
bug: image-scroll-up expects an integer, but was being given a float.  This
caused the call not to work as intended.  Former behavior would case the display
to "walk" up with repeated calls to `org-noter-sync-current-note'.

feature: defcustom `org-noter-vscroll-buffer' allows the user to customize how
far below the top of the window the page should be scrolled to.  Default 5,
while 0 gives what I believe was the intended behavior.