arrow positioning: move horizontal shift into --show-arrow
This is an attmept to make the arrow position less dependent on display
magnification.

Formerly, `org-noter-arrow-horizontal-offset' was in "percentage of page"
units.  The arrow, however, is displayed with constant display-pixels units, so
the needed offset as a percentage of the page was less for higher
magnification.  Also, the offset was encoded in the location of the note,
which could potentially affect placement of the note in multicolumn contexts
because the encoded position would be shifted from the click position.  Yet
another issue is that precise locations based on selected text was not being
treated the same way as clicked precise locations.

This commit changes `org-noter-arrow-horizontal-offset' to display-pixels units
and shifts the arrow in `org-noter--show-arrow', rather than shifting the
precise location in the `...-get-precise-info' function.