bugfix: issue #12 (open document to saved location)
With this change, if the notes file for a document (tested for PDFs only) has a
saved location in the top-level org entry, then upon starting the noter session,
the document display is moved to that page.

Previously, this only worked when the session was started from the notes file.

Other minor changes:

1. `t` is not documented as a value for the optional 4th
argument of `re-search-forward`.  Implicitly, it is acting as a `1`, but in the
documentation `nil` (the default) makes the 4th argument (# of searches forward)
`1`.  So the `t` is not needed.

2. Moby Dick in the "tests" subdir has as it's last saved location, page 171,
the earlier of the two "Knights and Squires" chapters.