override `C-c C-c` and `C-c C-x C-v` in pdf-view-mode
Takes keybindings that would not normally be used in pdf-view-mode buffer when
running org-noter and runs those keybindings in the notes buffer as if focus was
there all along.

`C-c C-c` is used by org-babel to run code.  If user has focus on the document,
then it would normally run `doc-view-toggle-display`, which takes the doc buffer
out of org-noter-doc-mode.

`C-c C-x C-v` is used by org-babel to toggle inline images.  If user has focus
on the document, then `C-c C-x` would normally run `image-toggle-hex-display`,
which takes the doc buffer out of org-noter-doc-mode.