Remove trailing "/" when expand load-path
Some distributions, such as Nixpkgs, use
`normal-top-level-add-subdirs-to-load-path' to recursively add sub
dirs of load-path entries to load-path.

org-note also does this when it is loaded using `add-to-list'.
Ideally, this should be a no-op since `add-to-list' does
deduplication.  However, the entry org-noter adds has a trailing "/",
which is not the case for other entries of `load-path'[1].  As a
result, the deduplication fails.

[1]: See the source code of `normal-top-level-add-to-load-path'.